Cast scale to the input_low/input_range dtype

+++ nncf/torch/quantization/quantize_functions.py
@@ -192,7 +192,8 @@ class TuneRange(torch.autograd.Function):
         input_low_copy[input_low_copy > 0] = 0
         input_high[input_high < 0] = 0
         n = levels - 1
-        scale = levels / (input_high - input_low_copy)
+        # Need a cast here because fp16 division yileds fp32 results sometimes
+        scale = (levels / (input_high - input_low_copy)).to(dtype=input_high.dtype)
         zp = torch.round(-input_low_copy * scale)
 
         new_input_low = torch.where(zp < n, zp / (zp - n) * input_high, input_low_copy)
